Correcion en clase, estar pendiente de los tipos datos

+++ Salario_Trabajador_D1.py
@@ -14,13 +14,13 @@ print("***PAGO DE NOMINA POR HORAS LABORADAS***") #TITULO DEL PROYECTO
 ##Ejemplo 1## valorxhora "30000"*horas_laboradas "35": debe de arrojar un salario de 1.050.000
 valorxhora = 30000
 horas_laboradas = 35
-sueldo_final =  int(valorxhora * horas_laboradas )
+sueldo_final =  float(valorxhora * horas_laboradas )
 print ("El sueldo final a devengar es de: ", sueldo_final, "ejemplo 1")
 
 ##Ejemplo 2## valorxhora "30000" * horas_laboradas "45": debe de arrojar un salario de 1350000
 valorxhora = 30000
 horas_laboradas = 45
-sueldo_final =  int(valorxhora * horas_laboradas )
+sueldo_final =  float(valorxhora * horas_laboradas )
 print ("El sueldo final a devengar es de: ", sueldo_final, "ejemplo 2")
 
 
@@ -38,7 +38,7 @@ print ("El sueldo final a devengar es de: ", sueldo_final, "ejemplo 2")
 
 valorxhora = 30000 #Valor definido por el enunciado
 horas_laboradas = float(input("Ingresa las horas laboradas: ")) # variable que encierra la entrada suministrada
-sueldo_final =  int(valorxhora * horas_laboradas ) # definicion que opera a las variables ya definidas
+sueldo_final =  float(valorxhora * horas_laboradas ) # definicion que opera a las variables ya definidas
 print ("El sueldo final a devengar es de: ", sueldo_final) # Llamado a la variable que muestra el resultado que buscamos
 
 # Encabezado
